split logging into the public API and the worker-process transport
